DCS-792 BVL fix session being retained after different user signs in (#125)
* DCS-792: Current user logged out if another user logs in.
Add data-qa attribute to target the display name of logged in user, which is in the blue bar at top of page

* DCS-792: integration tests updated
Integration test added to verify the logged in user changes when a new user logs in.
Changes made to stubLoginCourt() and stubUserMe() stub functions. User details passed to stubUserMe via stubLoginCourt.