DCS-866: removal of activeCaseloadId.  (#171)
* DCS-866: removal of activeCaseloadId.

This is not relevant to court users as they don't have a caseload. The application does still however need the prisoner's prison and this is now captured in the agencyId value. agencyId value has now been added to to the flash appointmentDetails so it can be accessed by the confirmationControllers.

* DCS-866-updated unit tests 

to verify agencyId placed into appointmentDetails and used in call to .getRooms().

* DCS-866 type reinstated following PR